chore(component): relocated & renamed wheel component
