--- packages/ui/src/components/wheel/ConfirmWheel.tsx
@@ -1,18 +1,0 @@
-'use client';
-
-import { Check, X } from 'lucide-react';
-
-import { Wheel } from './Wheel';
-
-function ConfirmWheel() {
-  return (
-    <Wheel
-      type="confirm"
-      icons={[Check, X]}
-      titles={['Confirm', 'Cancel']}
-      variant="destructive"
-    />
-  );
-}
-
-export default ConfirmWheel;
