--- packages/ui/src/components/wheel/index.ts
@@ -1,4 +1,0 @@
-export { default as ConfirmWheel } from './ConfirmWheel';
-export { default as FiveWheel } from './FiveWheel';
-export { default as FourWheel } from './FourWheel';
-export { Wheel } from './Wheel';
